pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "guardette"
version = "0.1.0"
authors = [
    { name="Span" }
]
description = "A redacting proxy built on FastAPI/Starlette."
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "annotated-types>=0.6.0",
    "anyio>=3.7.1",
    "anyio[trio]>=3.7.1",
    "aiobotocore>=2.7.0",
    "types-aiobotocore>=2.7.0",
    "types-aiobotocore[secretsmanager]>=2.7.0",
    "certifi>=2023.7.22",
    "click>=8.1.7",
    "cryptography>=41.0.5",
    "fastapi>=0.103.2",
    "h11>=0.14.0",
    "httpcore>=0.18.0",
    "httpx>=0.25.0",
    "idna>=3.4",
    # 1.6.1 .update() is broken
    "jsonpath-ng>=1.6.0,!=1.6.1",
    "ply>=3.11",
    "pydantic>=2.4.2",
    "pydantic_core>=2.10.1",
    "sniffio>=1.3.0",
    "starlette>=0.27.0",
    "typing_extensions>=4.8.0",
    "uvicorn>=0.23.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "PyJWT>=2.8.0",
    "pyjwt[crypto]>=2.8.0",
    "mangum>=0.17.0",
]

[project.optional-dependencies]
dev = ["build", "jinja2>=3.1.2", "pytest", "ruff", "requests"]

[project.urls]
Homepage = "https://github.com/attuned-corp/guardette"
Issues = "https://github.com/attuned-corp/guardette/issues"

[tool.pytest.ini_options]
pythonpath = ["src"]
